% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{order.group.mod}
\alias{order.group.mod}
\title{A modified version of order.group of agricolae package}
\usage{
order.group.mod(trt, means, N, MSerror, Tprob, std.err, parameter = 1,
  snk = 0, DFerror = NULL, alpha = NULL, sdtdif = NULL,
  vartau = NULL, console)
}
\arguments{
\item{trt}{Treatments}

\item{means}{Means of treatment}

\item{N}{Replications}

\item{MSerror}{Mean square error}

\item{Tprob}{Minimum value for the comparison}

\item{std.err}{Standard error}

\item{parameter}{Constante 1 (Sd), 0.5 (Sx)}

\item{snk}{Constante = 1 (Student Newman Keuls)}

\item{DFerror}{Degrees of freedom of the experimental error}

\item{alpha}{Level of risk for the test}

\item{sdtdif}{Standard deviation of difference in BIB}

\item{vartau}{Matrix var-cov in PBIB}

\item{console}{Logical, print output}
}
\description{
Extesion of the collection of letters and numbers for comparison means.
}
\author{
Omar Benites
}
