\encoding{UTF-8}

\name{predicted}
\alias{predicted}

\title{Sample Panels of Commodities Spot Prices Forecasts.}

\description{Forecasts obtained from various methods applied to various commodities prices.}

\usage{data(forecasts)}

\format{
\itemize{\code{predicted} is \code{\link[base]{list}} of forecasts of spot prices of selected 56 commodities. For each commodity \code{\link[base]{matrix}} of forecasts generated by various methods is provided. Columns correspond to various methods. 
}
}

\details{
The forecasts were taken from Drachal and Pawłowski (2024). They cover the period between 1996 and 2021, and are in monthly freqency. Variables and methods names are the same as in that paper, where they are described in details. 
}

\examples{
data(forecasts)
# WTI prices predicted by BSR rec method
t2 <- predicted[[3]][,1]
}

\references{
Drachal, K., Pawłowski, M. 2024. Forecasting selected commodities' prices with the Bayesian symbolic regression. \emph{International Journal of Financial Studies} \bold{12}, 34, \doi{10.3390/ijfs12020034}
}

\seealso{\code{\link[pEPA]{observed}}}
