% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pty_rcbd_withchild.R
\name{pty.rcbd.withchild}
\alias{pty.rcbd.withchild}
\title{Pepa tells you about RCBD}
\usage{
pty.rcbd.withchild(traits, treat, rep, data, maxp = 0.1,
  author = "International Potato Center")
}
\arguments{
\item{traits}{The traits to analize.}

\item{treat}{The treatments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{author}{Author.}
}
\value{
It returns an explanation about the RCBD fitted model.
}
\description{
Explain a RCBD fitted model in plain English
}
\details{
It fits a linear model for a RCBD and explains the results.
It also checks the assumptions.
}
\examples{
pty.rcbd.withchild(c("trw", "vw"), "geno", "rep", pjpz09)
}
\author{
Raul Eyzaguirre.
}

