% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pty.R
\name{pty}
\alias{pty}
\title{Pepa tells you}
\usage{
pty(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
It returns an automatic report about the selected R object.
}
\description{
Explain an R object in plain English if she knows about it
}
\details{
It uses a set of templates to explain R objects in plain English.
It aims to produce automatic reports for some standard statistical procedures, most
of them included in the \code{st4gi} package.
}
\examples{
# Pepa tells you something about a data frame:
pty(pjpz09)
}
\author{
Raul Eyzaguirre.
}

