% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pty_spconsis.R
\name{pty.spconsis}
\alias{pty.spconsis}
\title{Pepa tells you about the consistency of your data}
\usage{
pty.spconsis(data, plot.size, f = 3, author = "International Potato Center")
}
\arguments{
\item{data}{The name of the data frame.}

\item{plot.size}{Plot size in square meters.}

\item{f}{Factor for extreme values detection. See details.}

\item{author}{Author.}
}
\value{
It returns a list of all rows with some kind of inconsistency and
all rows with outliers.
}
\description{
It checks your data for inconsistencies.
}
\details{
The data frame must use the labels (lower or upper case) specified in
the function \code{spconsis} of package \code{st4gi}. Type \code{?spconsis} to
see the list.
Extreme values are detected using the interquartile range.
The rule is to detect any value out of the interval
\eqn{[Q_1 - f \times IQR; Q_3 + f \times IQR]}. By default \code{f = 3}.
}
\examples{
pty.spconsis(pjpz09, 4.5)
}
\author{
Raul Eyzaguirre.
}

