% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pty_rcbd.R
\name{pty_rcbd}
\alias{pty_rcbd}
\title{Pepa tells you about RCBD}
\usage{
pty_rcbd(trait, treat, block, data, author = "International Potato Center")
}
\arguments{
\item{trait}{The trait to analize.}

\item{treat}{The treatments.}

\item{block}{The blocks.}

\item{data}{The name of the data frame.}

\item{author}{Author.}
}
\value{
It returns an automatic report about the RCBD fitted model.
}
\description{
Explain a RCBD fitted model in plain English
}
\details{
It fits a linear model for a RCBD and explains the results.
It also checks the assumptions.
}
\examples{
pty_rcbd("trw", "geno", "rep", pjpz09)
}
\author{
Raul Eyzaguirre.
}

