% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pvs.R
\docType{data}
\name{pvsf1}
\alias{pvsf1}
\title{Voting for selection criteria}
\format{A data frame with 5 columns and 15 rows.}
\source{
International Potato Center, potato data.
}
\usage{
pvsf1
}
\description{
This data set has the voting of men (\code{SCORE_MEN}),
women (\code{SCORE_WOMEN}), and global voting (\code{SCORE_GLOBAL}) for
the more importan selection criteria (\code{IDENTIFIED_CRITERIA}) on
different phases (\code{PHASE}).
}

