% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pvsf2}
\alias{pvsf2}
\title{Voting for best clones at flowering}
\format{
A data frame with 8 columns and 18 rows.
}
\source{
International Potato Center, potato data.
}
\usage{
pvsf2
}
\description{
This data set has the voting of men (\code{MSM}), women (\code{MSWM}),
and global voting (\code{MSGLO}) in the mother plot, and the voting of men
(\code{BSM}), women (\code{BSWM}), and global voting (\code{BSGLO}) in the baby
plots for the best clones (\code{INSTN}) at flowering.
}
