% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pvsf6}
\alias{pvsf6}
\title{Organoleptic scores at mother trial}
\format{
A data frame with 6 columns and 50 rows.
}
\source{
International Potato Center, potato data.
}
\usage{
pvsf6
}
\description{
This data set has the scores given by men and women to several clones
(\code{INSTN}), for attributes \code{APPEARANCE}, \code{TASTE}, and \code{TEXTURE}
at the mother trial.
}
