% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pvsf9}
\alias{pvsf9}
\title{Voting for best clones at post-harvest}
\format{
A data frame with 6 columns and 18 rows.
}
\source{
International Potato Center, potato data.
}
\usage{
pvsf9
}
\description{
This data set has the voting of men (\code{SCORE_MEN}), women
(\code{SCORE_WOMEN}), and global voting (\code{SCORE_GLOBAL}) for the best
clones (\code{INSTN}) at post-harvest with different storage types (\code{STYPE}).
}
