% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pvssg}
\alias{pvssg}
\title{PVS summary global}
\format{
A data frame with 163 columns and 6 rows.
}
\source{
International Potato Center, potato data.
}
\usage{
pvssg
}
\description{
This data set has summaries (means, standard deviations) for several traits.
}
