% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_2f.R
\name{repo.2f}
\alias{repo.2f}
\title{Automatic report for a 2-factor factorial}
\usage{
repo.2f(traits, A, B, rep, design, data, maxp = 0.1,
  title = "Automatic report for a 2-factor factorial", subtitle = NULL,
  author = "International Potato Center", format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{A}{Factor A.}

\item{B}{Factor B.}

\item{rep}{The replications or blocks.}

\item{design}{The statistical design, \code{crd} or \code{rcbd}.}

\item{data}{The name of the data frame containing the data.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an automatic report about the 2-factor factorial with a CRD or
RCBD fitted model.
}
\description{
Produces an automatic report for selected traits in a 2-factor factorial with
a CRD or RCBD.
}
\details{
It fits a linear model for a 2-factor factorial with a CRD or RCBD for
the selected trait. If data is unbalanced, missing values are estimated up to an
specified maximum proportion, 10\% by default. Both factors are considered as fixed.
}
\examples{
repo.2f(c("asc.dw", "asc.fw"), "geno", "treat", "rep", "crd", asc)
}
\author{
Raul Eyzaguirre.
}
