% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_a01d.R
\name{repo.a01d}
\alias{repo.a01d}
\title{Automatic report for an alpha (0,1) design}
\usage{
repo.a01d(
  vars,
  geno,
  rep,
  block,
  k,
  method = c("REML", "VC"),
  dfr,
  title = "Automatic report for an alpha (0,1) design",
  subtitle = NULL,
  author = "International Potato Center",
  format = c("html", "word", "pdf"),
  server = FALSE,
  server_dir_name = "directory",
  server_file_name = "filename"
)
}
\arguments{
\item{vars}{The variables to analize.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{block}{The blocks.}

\item{k}{The size of the blocks.}

\item{method}{The estimation method, REML or VC (Variance Components).}

\item{dfr}{The name of the data frame.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an explanation about the alpha (0,1) design fitted model.
}
\description{
Produces an automatic report for selected variables in an experiment with an
alpha (0,1) design.
}
\details{
It fits a linear model for an alpha (0,1) design and explains the results.
An alpha (0,1) design is an incomplete block design that is resolvable.
In a resolvable design the incomplete blocks group together in complete replications.
}
\examples{
repo.a01d("yield", "geno", "rep", "block", 3, dfr = a01data)
}
\author{
Raul Eyzaguirre.
}
