% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_f.R
\name{repo.f}
\alias{repo.f}
\title{Automatic report for a factorial experiment}
\usage{
repo.f(traits, factors, rep, dfr, maxp = 0.1,
  title = "Automatic report for a factorial experiment",
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"), server = FALSE,
  server_dir_name = "directory", server_file_name = "filename")
}
\arguments{
\item{traits}{The traits to analize.}

\item{factors}{The factors.}

\item{rep}{The replications or blocks, \code{NULL} for a CRD.}

\item{dfr}{The name of the data frame containing the data.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an automatic report for the factorial experiment with a CRD or
RCBD fitted model.
}
\description{
Produces an automatic report for selected traits in a factorial experiment with
a CRD or RCBD.
}
\details{
It fits a linear model for a factorial experiment with a CRD or RCBD for
the selected traits. If data is unbalanced, missing values are estimated up to an
specified maximum proportion, 10\% by default. All factors are considered as fixed.
}
\examples{
## Example 1: Two factors factorial
repo.f(c("asc.dw", "asc.fw"), c("geno", "treat"), NULL, asc)

## Example 2: Three factors factorial
# Create design with 3 factors
fnames <- c('A', 'B', 'C')
flevels <- list(c('a1', 'a2'), c('b1', 'b2'), c('c1', 'c2', 'c3'))
temp <- cr.f(fnames, flevels, 'rcbd', 2, 10)$book
# Simulate random data
temp$y <- rnorm(24)
# Run report
repo.f('y', c('A', 'B', 'C'), 'block', temp)
}
\author{
Raul Eyzaguirre.
}
