% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_LxTs.R, R/repo_nc.R
\name{repo.LxT}
\alias{repo.LxT}
\alias{repo.LxT}
\title{Automatic report for a Line x Tester genetic design.}
\usage{
repo.LxT(traits, line, tester, rep, data,
  title = "Automatic report for a Line x Tester genetic design",
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"))

repo.LxT(traits, line, tester, rep, data,
  title = "Automatic report for a Line x Tester genetic design",
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{line}{The lines.}

\item{tester}{The testers.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{traits}{The traits to analize.}

\item{line}{The lines.}

\item{tester}{The testers.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an ANOVA and related quantities.

It returns an ANOVA and related quantities.
}
\description{
Produces an automatic report for a Line x Tester genetic design analysis.

Produces an automatic report for a Line x Tester genetic design analysis.
}
\details{
It fits a linear model to data from a Line x Tester genetic design.

It fits a linear model to data from a Line x Tester genetic design.
}
\examples{
library(agricolae)
data(LxT)
repo.LxT("yield", "line", "tester", "replication", LxT)
library(agricolae)
data(LxT)
repo.LxT("yield", "line", "tester", "replication", LxT)
}
\author{
Raul Eyzaguirre.

Raul Eyzaguirre.
}
