% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_met.R
\name{repo.met}
\alias{repo.met}
\title{Automatic report for a MET with a RCBD}
\usage{
repo.met(traits, geno, env, rep, data, maxp = 0.1,
  title = "Automatic report for a MET with a RCBD", subtitle = NULL,
  author = "International Potato Center", format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame containing the data.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an automatic report about the MET with a RCBD fitted model.
}
\description{
Produces an automatic report for selected traits in a multi environment
trial (MET) with a RCBD in each environment.
}
\details{
It fits a linear model for a MET with a RCBD for the selected trait.
If data is unbalanced, missing values are estimated up to an specified maximum
proportion, 10\% by default. Genotypes and environments are considered as fixed
factors while the blocks are considered as random and nested into the environments
for ANOVA. For variance components estimation all the factors
are treated as random.
}
\examples{
repo.met(c("rytha", "fytha"), "geno", "env", "rep", megaclones)
}
\author{
Raul Eyzaguirre.
}

