% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pvs_stage.R
\name{repo.pvs.stage}
\alias{repo.pvs.stage}
\title{Report for PVS at stage}
\usage{
repo.pvs.stage(
  dfr,
  stage = NULL,
  geno = "geno",
  rep = "rep",
  msm = "msm",
  msw = "msw",
  msg = "msg",
  bsm = "bsm",
  bsw = "bsw",
  bsg = "bsg",
  title = "Automatic report for PVS",
  subtitle = "Best genotypes at",
  author = "International Potato Center",
  format = c("html", "word", "pdf")
)
}
\arguments{
\item{dfr}{A data frame with data for a pvs at some stage.}

\item{stage}{\code{flowering}, \code{harvest} or \code{post-harvest}.
This is only for titles in the report.}

\item{geno}{Name of the genotypes.}

\item{rep}{Replications.}

\item{msm}{Mother score for men.}

\item{msw}{Mother score for women.}

\item{msg}{Mother score global.}

\item{bsm}{Baby score for men.}

\item{bsw}{Baby score for women.}

\item{bsg}{Baby score global.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an automatic report for the PVS data at different stages:
flowering, harvest, posthavest.
}
\description{
Produce standard reports for PVS data at different stages.
}
\examples{
repo.pvs.stage(pvsf2, 'flowering', 'INSTN', 'REP',
               'MSM', 'MSWM', 'MSGLO', 'BSM', 'BSWM', 'BSGLO')
}
\author{
Raul Eyzaguirre.
}
