% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pvssg.R
\name{repo.pvssg}
\alias{repo.pvssg}
\title{Report for PVS - Global summary}
\usage{
repo.pvssg(traits, data, title = "Automatic report for PVS",
  subtitle = "Global summary data", author = "International Potato Center",
  format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{A list of traits to be included in the report.}

\item{data}{A data frame with data for a pvs global summary form.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an automatic report.
}
\description{
Produce a report for PVS data global summary sheet
}
\details{
It produces a report for global summary sheet of the Mother & Baby methodology
implemented by CIP.
}
\examples{
traits <- c("NMTP_Mean_mother", "TNTP_Mean_mother", "PPH_Mean_mother", "TTYA_Mean_mother")
repo.pvssg(traits, pvssg)
}
\author{
Raul Eyzaguirre.
}

