% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_rcbd.R
\name{repo.rcbd}
\alias{repo.rcbd}
\title{Authomatic report for a Randomized Complete Block Design (RCBD)}
\usage{
repo.rcbd(traits, treat, rep, data, maxp = 0.1,
  author = "International Potato Center")
}
\arguments{
\item{traits}{The traits to analize.}

\item{treat}{The treatments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{author}{Author.}
}
\value{
It returns an explanation about the RCBD fitted model.
}
\description{
Produces an authomatic report for selected traits in an experiment with a RCBD.
}
\details{
It fits a linear model for a RCBD and explains the results.

Under the assumption of fixed effects an ANOVA table is computed with missing values
estimated up to a specified percentage (10\% by default). If the ANOVA results in
a significant value for treatments then the Tukey HSD method for pairwise differences
is applied. Assumptions of the model are evaluated with residual plots.

Under the assumption of random effects the model is estimated using REML and the
variance components are shown. Missing values are not estimated in this case.
}
\examples{
repo.rcbd(c("trw", "vw", "crw"), "geno", "rep", pjpz09)

# With a small data set
temp <- pjpz09[1:18, ]
repo.rcbd(c("trw", "vw", "crw"), "geno", "rep", temp)
}
\author{
Raul Eyzaguirre.
}

