% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_spld.R
\name{repo.spld}
\alias{repo.spld}
\title{Automatic report for a split plot design}
\usage{
repo.spld(traits, mpf, spf, sspf = NULL, rep, dfr, title = NULL,
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"), server = FALSE,
  server_dir_name = "directory", server_file_name = "filename")
}
\arguments{
\item{traits}{The traits to analize.}

\item{mpf}{The main plots factor.}

\item{spf}{The sub plots factor.}

\item{sspf}{The sub sub plot factor (if any).}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an explanation about the split-plot or a split-split-plot design
fitted model.
}
\description{
Produces an automatic report for selected traits in an experiment
with a split-plot or a split-split-plot design.
}
\details{
It fits a linear model for a split-plot or a split-split-plot design and explains
the results.
}
\examples{
# Create designs
mpf <- paste0("mpf", 1:2)
spf <- paste0("spf", 1:3)
sspf <- paste0("sspf", 1:4)
d1 <- cr.spld(c("mpf", "spf"), list(mpf, spf), 3)$book
d2 <- cr.spld(c("mpf", "spf", "sspf"), list(mpf, spf, sspf), 2)$book
# Some random data
d1$yield <- rnorm(18, 10)
d2$yield <- rnorm(48, 10)
# Fit models
repo.spld("yield", "mpf", "spf", NULL, "block", d1)
repo.spld("yield", "mpf", "spf", "sspf", "block", d2)
}
\author{
Raul Eyzaguirre.
}
