% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_strd.R
\name{repo.strd}
\alias{repo.strd}
\title{Automatic report for a strip plot design}
\usage{
repo.strd(
  dfr,
  vars,
  rowf,
  colf,
  rep,
  title = "Automatic report for a strip plot design",
  subtitle = NULL,
  author = "International Potato Center",
  format = c("html", "word", "pdf"),
  server = FALSE,
  server_dir_name = "directory",
  server_file_name = "filename"
)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{vars}{The variables to analize.}

\item{rowf}{The row factor.}

\item{colf}{The column factor.}

\item{rep}{The replications.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an explanation about the strip plot design fitted model.
}
\description{
Produces an automatic report for selected variables in an experiment
with a strip plot design.
}
\details{
It fits a linear model for a strip plot design and explains the results.
}
\examples{
# Create design
rowf <- paste0("rf", 1:3)
colf <- paste0("cf", 1:4)
design <- cr.strd(rowf, colf, 3)$book
# Some random data
design$yield <- rnorm(36, 10)
repo.strd(design, "yield", "A", "B", "block")
}
\author{
Raul Eyzaguirre.
}
