% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_trend.R
\name{repo.trend}
\alias{repo.trend}
\title{Report for repeated measures data}
\usage{
repo.trend(response, treat, eu, dap, dfr, average = TRUE, se = TRUE,
  title = "Automatic report for repeated measures data",
  subtitle = "Dispersion plot with a fitted smoothing function",
  author = "International Potato Center", format = c("html", "word",
  "pdf"), server = FALSE, server_dir_name = "directory",
  server_file_name = "filename")
}
\arguments{
\item{response}{Response variable.}

\item{treat}{The treatments.}

\item{eu}{The experimental unit. Must be defined in case of subsamples.}

\item{dap}{Days after planting measuring time.}

\item{dfr}{The name of the data frame.}

\item{average}{If \code{"TRUE"}, it computes average over experimental units
(for subsampling data).}

\item{se}{If \code{"TRUE"}, standard errors are ploted.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns a ggplot2 graph for repeated measures data.
}
\description{
Produce a ggplot2 graph for data over time.
}
\examples{
repo.trend("yield", "treat", "plot", "date", trend.data)
repo.trend("yield", "treat", "plot", "date", trend.data, FALSE)
repo.trend("yield", "treat", "plot", "date", trend.data, FALSE, FALSE)
}
\author{
Raul Eyzaguirre.
}
