% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trend.data}
\alias{trend.data}
\title{Repeated measures data}
\format{A data frame with 8 columns and 32 rows.}
\source{
Dummy data.
}
\usage{
trend.data
}
\description{
Repeated measures data for trait yield.
}
