% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pty_elston.R
\name{pty.elston}
\alias{pty.elston}
\title{Pepa tells you about the Elston index}
\usage{
pty.elston(traits, geno, env = NULL, rep = NULL, data, means = "single",
  model = "gxe", lb = 1, author = "International Potato Center")
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications. Must be defined if \code{model = "gxe"}.}

\item{data}{The name of the data frame containing the data.}

\item{means}{The genotypic means to compute the index, \code{"single"}
or \code{"fitted"}. The default is \code{"single"}. See details for more information.}

\item{model}{Type of model to fit means if \code{means = "fitted"}, \code{"gxe"} for
a model with gxe interaction or \code{"g+e"} for a model without interaction.
The default is \code{"gxe"}. See details for more information.}

\item{lb}{Lower bound. \code{1} for \eqn{k = min(x)} and \code{2} for
\eqn{k = (n \times min(x) - max(x)) / (n - 1)}}

\item{author}{Author.}
}
\value{
It returns an explanation about the Elston index.
}
\description{
Explain the results of the Elston index in plain English.
}
\details{
The Elston index is a weight free index.

If \code{means = "fitted"} and \code{model = "gxe"} then the arguments \code{env} and
\code{rep} must be specified.
If \code{means = "fitted"} and \code{model = "g+e"} then only the argument \code{env}
must be specified.
If \code{means = "single"} and \code{env} and \code{rep} are specified, then
single arithmetic means are computed over the replications for each genotype
at each environment and then for each genotype over environments. In any other case
single arithmetic means are computed over all the observations for each genotype.
}
\examples{
pty.elston(c("rytha", "bc", "dm", "star", "nocr"), "geno", data = spg)
}
\author{
Raul Eyzaguirre.
}

