% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pty_pesekbaker.R
\name{pty.pesekbaker}
\alias{pty.pesekbaker}
\title{Pepa tells you about the Pesek-Baker index}
\usage{
pty.pesekbaker(traits, geno, env, rep = NULL, dfr, means = "single",
  dgg = NULL, units = "sdu", sf = 0.1,
  author = "International Potato Center", format = c("html", "word",
  "pdf"))
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications. Must be defined if \code{model = "gxe"}.}

\item{dfr}{The name of the data frame containing the data.}

\item{means}{The genotypic means to compute the index, \code{"single"} or
\code{"fitted"}. The default is \code{"single"}. See details for more information.}

\item{dgg}{Desired genetic gains. The default is one standard deviation for each trait.}

\item{units}{Units for dgg, \code{"actual"} or \code{"sdu"}.
See details for more information.}

\item{sf}{Selected fraction. The default is 0.1.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an explanation about the Pesek-Baker index.
}
\description{
Explain the results of the Pesek-Baker index in plain English.
}
\details{
Type \code{?pesekbaker} for additional details.
}
\examples{
pty.pesekbaker(c("rytha", "bc", "dm", "star", "nocr"), "geno", "loc", "rep", spg)
}
\author{
Raul Eyzaguirre.
}
