% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pty.R
\name{pty}
\alias{pty}
\title{Pepa tells you}
\usage{
pty(x, author = "International Potato Center", format = c("html",
  "word", "pdf"))
}
\arguments{
\item{x}{An R object.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an explanation about the selected R object.
}
\description{
Explain an R object in plain English if she knows about it.
}
\details{
It uses a set of templates to explain R objects in plain English.
It aims to produce automatic reports for some standard statistical procedures, most
of them included in the \code{st4gi} package.
}
\examples{
# Pepa tells you something about a data frame:
pty(pjpz09)
}
\author{
Raul Eyzaguirre.
}
