% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_a01d.R
\name{repo.a01d}
\alias{repo.a01d}
\title{Automatic report for an alpha (0,1) design.}
\usage{
repo.a01d(traits, geno, rep, block, k, data,
  title = "Automatic report for an alpha (0,1) design", subtitle = NULL,
  author = "International Potato Center", format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{block}{The blocks.}

\item{k}{The size of the blocks.}

\item{data}{The name of the data frame.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an explanation about the alpha (0,1) design fitted model.
}
\description{
Produces an automatic report for selected traits in an experiment with an
alpha (0,1) design.
}
\details{
It fits a linear model for an alpha (0,1) design and explains the results.
An alpha (0,1) design is an incomplete block design that is resolvable.
In a resolvable design the incomplete blocks group together in complete replications.
}
\examples{
repo.a01d("yield", "geno", "rep", "block", k = 3, alpha01data)
}
\author{
Raul Eyzaguirre.
}
