% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_abd.R
\name{repo.abd}
\alias{repo.abd}
\title{Automatic report for an augmented block design (ABD)}
\usage{
repo.abd(
  dfr,
  vars,
  geno,
  rep,
  title = "Automatic report for an Augmented Block Design (ABD)",
  subtitle = NULL,
  author = "International Potato Center",
  format = c("html", "word", "pdf"),
  server = FALSE,
  server_dir_name = "directory",
  server_file_name = "filename"
)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{vars}{The variables to analize.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an explanation about the ABD fitted model.
}
\description{
Produces an automatic report for selected variables in an experiment with an ABD.
}
\details{
It fits a linear model for an ABD and explains the results.
}
\examples{
# A subset that looks like an ABD
tmp <- pjpz09[c(1, 2, 9, 10, 13, 14, 27, 29, 31, 33, 35, 37, 40, 42, 44, 46, 48, 50, 169, 170), ]
repo.abd(tmp, c("trw", "vw"), "geno", "rep")
# With some missing values
tmp[c(1, 2, 3), "trw"] <- NA
tmp[c(1, 10, 15), "vw"] <- NA
repo.abd(tmp, c("nocr", "trw", "vw"), "geno", "rep")
}
\author{
Raul Eyzaguirre.
}
