% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_abd.R
\name{repo.abd}
\alias{repo.abd}
\title{Automatic report for an augmented block design (ABD)}
\usage{
repo.abd(traits, geno, rep, data,
  title = "Automatic report for an Augmented Block Design (ABD)",
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an explanation about the ABD fitted model.
}
\description{
Produces an automatic report for selected traits in an experiment with an ABD.
}
\details{
It fits a linear model for an ABD and explains the results.
}
\examples{
# A subset that looks like an ABD
temp <- pjpz09[c(1, 2, 9, 10, 13, 14, 27, 29, 31, 33, 35, 37, 40, 42, 44, 46, 48, 50, 203, 204), ]
repo.abd(c("trw", "vw"), "geno", "rep", temp)

# With some missing values
temp[c(1, 2, 3), "trw"] <- NA
temp[c(1, 10, 15), "vw"] <- NA
repo.abd(c("nocr", "trw", "vw"), "geno", "rep", temp)
}
\author{
Raul Eyzaguirre.
}
