% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_abd.R
\name{repo.abd}
\alias{repo.abd}
\title{Authomatic report for an augmented block design (ABD)}
\usage{
repo.abd(traits, treat, rep, data, author = "International Potato Center")
}
\arguments{
\item{traits}{The traits to analize.}

\item{treat}{The treatments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{author}{Author.}
}
\value{
It returns an explanation about the ABD fitted model.
}
\description{
Produces an authomatic report for selected traits in an experiment with an ABD.
}
\details{
It fits a linear model for an ABD and explains the results.

The Tukey HSD method for pairwise differences is applied.
Assumptions of the model are evaluated with residual plots.
}
\examples{
# A subset that looks like an ABD
temp <- pjpz09[c(1, 2, 9, 10, 13, 14, 27, 29, 31, 33, 35, 37, 40, 42, 44, 46, 48, 50), ]
repo.abd(c("trw", "vw", "crw"), "geno", "rep", temp)
}
\author{
Raul Eyzaguirre.
}

