% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_crd.R
\name{repo.crd}
\alias{repo.crd}
\title{Automatic report for a Completely Randomized Design (CRD)}
\usage{
repo.crd(traits, geno, data, maxp = 0.1,
  title = "Automatic report for a Completely Randomized Design (CRD)",
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{geno}{The genotypes.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an explanation about the CRD fitted model.
}
\description{
Produces an automatic report for selected traits in an experiment with a CRD.
}
\details{
It fits a linear model for a CRD and explains the results.

Under the assumption of fixed effects an ANOVA table is computed. If the ANOVA
results in a significant value then the Tukey HSD method for pairwise differences
is applied. Assumptions of the model are evaluated with residual plots.

Under the assumption of random effects the model is estimated using REML and the
variance components are shown.
}
\examples{
repo.crd(c("trw", "vw"), "geno", pjpz09)

# With a small data set
temp <- pjpz09[1:18, ]
repo.crd(c("trw", "vw", "crw"), "geno", temp)
}
\author{
Raul Eyzaguirre.
}

