% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_f.R
\name{repo.f}
\alias{repo.f}
\title{Automatic report for a factorial experiment}
\usage{
repo.f(
  dfr,
  vars,
  factors,
  rep = NULL,
  maxp = 0.1,
  pe = FALSE,
  se = FALSE,
  title = "Automatic report for a factorial experiment",
  subtitle = NULL,
  author = "International Potato Center",
  format = c("html", "word", "pdf")
)
}
\arguments{
\item{dfr}{The name of the data frame containing the data.}

\item{vars}{The variables to analize.}

\item{factors}{The factors.}

\item{rep}{The replications or blocks, default is \code{NULL} for a CRD.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{pe}{Logical. If \code{"pe = TRUE"} multiple comparison tests for principal effects
are included even if interaction is significat, only in the case of 2 factors.
Default to \code{"pe = FALSE"}.}

\item{se}{Logical. If \code{"se = TRUE"} multiple comparison tests for simple effects
are included even if interaction is not significat, only in the case of 2 factors.
Default to \code{"se = FALSE"}.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an automatic report for the factorial experiment with a CRD or
RCBD fitted model.
}
\description{
Produces an automatic report for selected variables in a factorial experiment with
a CRD or RCBD.
}
\details{
It fits a linear model for a factorial experiment with a CRD or RCBD for
the selected variables. If data is unbalanced, missing values are estimated up to an
specified maximum proportion, 10\% by default. All factors are considered as fixed.
}
\examples{
## Example 1: Two factors factorial
repo.f(asc, c("asc.dw", "asc.fw"), c("geno", "treat"))

## Example 2: Three factors factorial
# Create design with 3 factors
fnames <- c('A', 'B', 'C')
flevels <- list(c('a1', 'a2'), c('b1', 'b2'), c('c1', 'c2', 'c3'))
tmp <- cr.f(fnames, flevels, 'rcbd', 2, 10)$book
# Simulate random data
tmp$y <- rnorm(24)
# Run report
repo.f(tmp, 'y', c('A', 'B', 'C'), 'block')
}
\author{
Raul Eyzaguirre.
}
