% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_lxt.R
\name{repo.lxt}
\alias{repo.lxt}
\title{Automatic report for a Line x Tester genetic design.}
\usage{
repo.lxt(traits, lines, testers, rep, data,
  title = "Automatic report for a Line x Tester genetic design",
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{lines}{The lines.}

\item{testers}{The testers.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an ANOVA and related quantities.
}
\description{
Produces an automatic report for a Line x Tester genetic design analysis.
}
\details{
It fits a linear model to data from a Line x Tester genetic design.
}
\examples{
repo.lxt("yield", "line", "tester", "rep", lxt)
}
\author{
Raul Eyzaguirre.
}
