% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_met.R
\name{repo.met}
\alias{repo.met}
\title{Automatic report for a MET with a RCBD}
\usage{
repo.met(
  traits,
  geno,
  env,
  rep,
  dfr,
  maxp = 0.1,
  pe = FALSE,
  se = FALSE,
  title = "Automatic report for a MET with a RCBD",
  subtitle = NULL,
  author = "International Potato Center",
  format = c("html", "word", "pdf"),
  server = FALSE,
  server_dir_name = "directory",
  server_file_name = "filename"
)
}
\arguments{
\item{traits}{The traits to analize.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame containing the data.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{pe}{Logical. If \code{"pe = TRUE"} multiple comparison tests for principal effects
are included even if interaction is significat. Default to \code{"pe = FALSE"}.}

\item{se}{Logical. If \code{"se = TRUE"} multiple comparison tests for simple effects
are included even if interaction is not significat. Default to \code{"se = FALSE"}.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an automatic report about the MET with a RCBD fitted model.
}
\description{
Produces an automatic report for selected traits in a multi environment
trial (MET) with a RCBD in each environment.
}
\details{
It fits a linear model for a MET with a RCBD for the selected trait.
If data is unbalanced, missing values are estimated up to an specified maximum
proportion, 10\% by default. Genotypes and environments are considered as fixed
factors while the blocks are considered as random and nested into the environments
for ANOVA. For variance components estimation all the factors are treated as random.
}
\examples{
repo.met(c("rytha", "fytha"), "geno", "env", "rep", megaclones)
}
\author{
Raul Eyzaguirre.
}
