% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_nc.R
\name{repo.nc}
\alias{repo.nc}
\title{Automatic report for a North Carolina genetic design.}
\usage{
repo.nc(traits, set, male, female, progeny, rep, model = NULL, data,
  title = "Automatic report for a North Carolina genetic design",
  subtitle = NULL, author = "International Potato Center",
  format = c("html", "word", "pdf"))
}
\arguments{
\item{traits}{The traits to analize.}

\item{set}{The set.}

\item{male}{The male.}

\item{female}{The female.}

\item{progeny}{The progeny (only for Carolina I design).}

\item{rep}{The replication.}

\item{model}{1, 2 or 3.}

\item{data}{The data frame.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an ANOVA and related quantities.
}
\description{
Produces an automatic report for a North Carolina I, II or III genetic design analysis.
}
\details{
It fits a linear model to data from a North Carolina I, II or III
genetic design.
}
\examples{
library(agricolae)
data(DC)
carolina1 <- DC$carolina1
repo.nc("yield", "set", "male", "female", "progenie", "rep", 1, carolina1)
}
\author{
Raul Eyzaguirre.
}
