% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_nc.R
\name{repo.nc}
\alias{repo.nc}
\title{Automatic report for a North Carolina genetic design.}
\usage{
repo.nc(traits, set, male, female, progeny = NULL, rep, model, dfr,
  title = NULL, subtitle = NULL,
  author = "International Potato Center", format = c("html", "word",
  "pdf"), server = FALSE, server_dir_name = "directory",
  server_file_name = "filename")
}
\arguments{
\item{traits}{The traits to analize.}

\item{set}{The set.}

\item{male}{The male.}

\item{female}{The female.}

\item{progeny}{The progeny (only for Carolina I design).}

\item{rep}{The replication.}

\item{model}{1, 2 or 3.}

\item{dfr}{The data frame.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an ANOVA and related quantities.
}
\description{
Produces an automatic report for a North Carolina I, II or III genetic design analysis.
}
\examples{
library(agricolae)
data(DC)
carolina1 <- DC$carolina1
repo.nc("yield", "set", "male", "female", "progenie", "rep", 1, carolina1)
carolina2 <- DC$carolina2
carolina2 <- carolina2[carolina2$Loc == 1, c(2, 5, 4, 3, 7, 8)]
repo.nc(c("yield", "tuber"), "set", "male", "female", NULL, "rep", 2, carolina2)
carolina3 <- DC$carolina3
repo.nc("yield", "set", "male", "female", NULL, "rep", 3, carolina3)
}
\author{
Raul Eyzaguirre.
}
