% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pvs.R
\name{repo.pvs}
\alias{repo.pvs}
\title{Report for PVS}
\usage{
repo.pvs(data, form, title = "Automatic report for PVS", subtitle = NULL,
  author = "International Potato Center", format = c("html", "word", "pdf"))
}
\arguments{
\item{data}{A data frame with data for a pvs form.}

\item{form}{Form number.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns automatic reports for the different forms on a PVS book.
}
\description{
Produce standard reports for PVS data
}
\details{
It produces a report for different forms of the Mother & Baby methodology
implemented by CIP.
}
\examples{
repo.pvs(pvsf1, 1)
repo.pvs(pvsf2, 2)
repo.pvs(pvsf3, 3)
repo.pvs(pvsf6, 6)
repo.pvs(pvsf7, 7)
repo.pvs(pvsf9, 9)
}
\author{
Raul Eyzaguirre.
}

