% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pvs.R
\name{repo.pvs}
\alias{repo.pvs}
\title{Report for PVS}
\usage{
repo.pvs(x, form, title = "Automatic report for PVS", subtitle = NULL,
  author = "International Potato Center", format = c("html", "word", "pdf"))
}
\arguments{
\item{x}{A data frame with data for a pvs form.}

\item{form}{Form number.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{author}{Author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns automatic reports for the different forms on a PVS book.
}
\description{
Produce standard reports for PVS data
}
\details{
bla bla bla.
}
\examples{
# bla bla bla
repo.pvs(pvsf1, 1)
}
\author{
Raul Eyzaguirre.
}

