% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_pvssg.R
\name{repo.pvssg}
\alias{repo.pvssg}
\title{Report for PVS - Global summary}
\usage{
repo.pvssg(
  traits,
  dfr,
  title = "Automatic report for PVS",
  subtitle = "Global summary data",
  author = "International Potato Center",
  format = c("html", "word", "pdf"),
  server = FALSE,
  server_dir_name = "directory",
  server_file_name = "filename"
)
}
\arguments{
\item{traits}{A list of traits to be included in the report.}

\item{dfr}{A data frame with data for a pvs global summary form.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an automatic report.
}
\description{
Produce a report for PVS data global summary sheet.
}
\details{
It produces a report for global summary sheet of the Mother & Baby methodology
implemented by CIP.
}
\examples{
traits <- c("NMTP_Mean_mother", "TNTP_Mean_mother", "PPH_Mean_mother", "TTYA_Mean_mother")
repo.pvssg(traits, pvssg)
}
\author{
Raul Eyzaguirre.
}
