% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_rcbd.R
\name{repo.rcbd}
\alias{repo.rcbd}
\title{Authomatic report for a RCBD}
\usage{
repo.rcbd(traits, treat, rep, data, maxp = 0.1,
  author = "International Potato Center")
}
\arguments{
\item{traits}{The traits to analize.}

\item{treat}{The treatments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{author}{Author.}
}
\value{
It returns an explanation about the RCBD fitted model.
}
\description{
Produces an authomatic report for selected traits in an experiment with a RCBD.
}
\details{
It fits a linear model for a RCBD and explains the results.
It also checks the assumptions.
}
\examples{
repo.rcbd(c("trw", "vw", "crw"), "geno", "rep", pjpz09)
}
\author{
Raul Eyzaguirre.
}

