% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_rcbd.R
\name{repo.rcbd}
\alias{repo.rcbd}
\title{Automatic report for a Randomized Complete Block Design (RCBD)}
\usage{
repo.rcbd(
  dfr,
  vars,
  trt,
  trt.lab = "treatment",
  rep,
  eu = NULL,
  maxp = 0.1,
  mc = FALSE,
  title = "Automatic report for a Randomized Complete Block Design (RCBD)",
  subtitle = NULL,
  author = "International Potato Center",
  format = c("html", "word", "pdf")
)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{vars}{The variables to analize.}

\item{trt}{The treatments.}

\item{trt.lab}{The labels for treatments.}

\item{rep}{The replications.}

\item{eu}{The experimental unit. Must be defined in case of subsamples.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{mc}{Logical. If \code{"mc = TRUE"} multiple comparison tests are included
even if the factor effect is not significat. Default to \code{"mc = FALSE"}.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}
}
\value{
It returns an explanation about the RCBD fitted model.
}
\description{
Produces an automatic report for selected variables in an experiment with a RCBD.
}
\details{
It fits a linear model for a RCBD and explains the results.

Under the assumption of fixed effects an ANOVA table is computed with missing
values estimated up to a specified percentage (10\% by default).
In case of subsampling the ANOVA table is computed over the means for each
experimental unit, and a second ANOVA table is computed considering subsampling
only if there are no missing values.
If the ANOVA results in a significant value for treatments then the Tukey HSD
method for pairwise differences is applied. Assumptions of the model are evaluated
with residual plots.

Under the assumption of random effects the model is estimated using REML and the
variance components are shown. Missing values are not estimated in this case.
}
\examples{
repo.rcbd(pjpz09, c("trw", "vw"), "geno", "genotype", "rep")
# With a small data set
tmp <- pjpz09[1:18, ]
repo.rcbd(tmp, c("trw", "vw", "crw"), "geno", "genotype", "rep")
}
\author{
Raul Eyzaguirre.
}
