% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_rcbd.R
\name{repo.rcbd}
\alias{repo.rcbd}
\title{Automatic report for a Randomized Complete Block Design (RCBD)}
\usage{
repo.rcbd(
  traits,
  trt,
  trt.lab = "treatment",
  rep,
  eu = NULL,
  dfr,
  maxp = 0.1,
  title = "Automatic report for a Randomized Complete Block Design (RCBD)",
  subtitle = NULL,
  author = "International Potato Center",
  format = c("html", "word", "pdf"),
  server = FALSE,
  server_dir_name = "directory",
  server_file_name = "filename"
)
}
\arguments{
\item{traits}{The traits to analize.}

\item{trt}{The treatments.}

\item{trt.lab}{The labels for treatments.}

\item{rep}{The replications.}

\item{eu}{The experimental unit. Must be defined in case of subsamples.}

\item{dfr}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{title}{Report title.}

\item{subtitle}{Report subtitle.}

\item{author}{Report author.}

\item{format}{The output file format for the report, \code{"html"} by default.
Other options are \code{"word"} and \code{"pdf"}.}

\item{server}{Logical. If \code{"FALSE"} works with local machines.
Otherwise works in server environments.}

\item{server_dir_name}{If \code{"server = TRUE"}, this is the directory name in the server.}

\item{server_file_name}{If \code{"server = TRUE"}, this is the file name in the server.}
}
\value{
It returns an explanation about the RCBD fitted model.
}
\description{
Produces an automatic report for selected traits in an experiment with a RCBD.
}
\details{
It fits a linear model for a RCBD and explains the results.

Under the assumption of fixed effects an ANOVA table is computed with missing
values estimated up to a specified percentage (10\% by default).
In case of subsampling the ANOVA table is computed over the means for each
experimental unit, and a second ANOVA table is computed considering subsampling
only if there are no missing values.
If the ANOVA results in a significant value for treatments then the Tukey HSD
method for pairwise differences is applied. Assumptions of the model are evaluated
with residual plots.

Under the assumption of random effects the model is estimated using REML and the
variance components are shown. Missing values are not estimated in this case.
}
\examples{
repo.rcbd(c("trw", "vw"), "geno", "genotype", "rep", dfr = pjpz09)
# With a small data set
temp <- pjpz09[1:18, ]
repo.rcbd(c("trw", "vw", "crw"), "geno", "genotype", "rep", dfr = temp)
}
\author{
Raul Eyzaguirre.
}
